backend/requirements.txt
fastapi
pydantic
openai
langchain
faiss-cpu
dotenv
python-dotenv
pydantic-ai
langgraph
